React Type Speed Test Ap
Today i am building an application for the user to test their typing speed by typing out words and having the speed and accuracy of this displayed for the user to see their results. In this first commit i have laid the foundation and gotten what i felt was going to be the most challenging part which was getting the text to work as intended which is now (almost) complete.