Functionality Implementation
Big updates to the overall functionality. I have implemented the functionality for the word tracker which is dynamic to the amount of words given to the user.

I have fixed issues with too many spaces being typed resulting in the total count being off by preventing the user from typing consecutive spaces.

I have implemented the timer functionality and unmounted the component once the timer has ended.

The results component is the newest edition and it will be the component rendered in place of the text after the timer has ended. This is what i will be working on this week to fully finish.