Finishing Touches
In this update i have made a lot of good progress. I have added functionality for the timer and moved it to its own component and given the reset button the ability to manage the timer so that it is restarted after the user begins typing on the new text.

I have also added a very simplistic results page to display the WPM or words per minute of the user as well as their accuracy.

Finally i have added some further minimalistic styling to the header and main elements.

This will likely be version 1 with version 2 to come once i am further along with react. There are a few ways i feel i should do things differently as well as wanting to add a graph component to show the results using rechart and framer motion. However it is clear i will things like custom hooks and other react hooks that i am not yet familar with such as useContext etc.